requirements.txt
easydict==1.10
matplotlib==3.6.2
mlflow==2.0.1
numpy==1.23.5
pandas==1.5.2
pydantic==1.10.2
PyYAML==6.0
scikit_learn==1.2.0
seaborn==0.12.1
torch==1.13.0+cu117
torchinfo==1.7.1
torchvision==0.14.0+cu117
tqdm==4.64.1
